split jekyll.spec.ts to small ones
