+++ tests/jekyll/front-matter-image.spec.ts
@@ -1,0 +1,43 @@
+#!/usr/bin/env ts-node
+
+import test  from 'tstest'
+
+import util   from 'util'
+import fs     from 'fs'
+import globCB from 'glob'
+
+import { loadFront } from 'yaml-front-matter'
+
+import {
+  getYearMonth,
+  JEKYLL_FOLDER,
+}                             from '../../src/jekyll/mod'
+
+import {
+  stripRepoRoot,
+}                             from '../../src/repo-root'
+
+const glob = util.promisify(globCB)
+
+/**
+ * Issue #351: https://github.com/wechaty/wechaty.js.org/issues/351
+ *  Should add teaser for the blog
+ */
+test('front matter key `image` must has a value to define the teaser image', async t => {
+  const postsFileList   = await glob(`${JEKYLL_FOLDER.posts}/**/*`)
+
+  for (const file of postsFileList) {
+    const { year } = getYearMonth(file)
+    /**
+     * Huan(202101): We leave the posts before 2021 as it is
+     */
+    if (parseInt(year) < 2021) {
+      continue
+    }
+
+    const content = fs.readFileSync(file)
+    const front = loadFront(content)
+    const image = front.image
+    t.true(image, `"${stripRepoRoot(file)}" image(${image}) has been set`)
+  }
+})
